+++ x/crosschain/types/expected_keepers.go
@@ -50,7 +50,7 @@ type ObserverKeeper interface {
 	GetBallot(ctx sdk.Context, index string) (val observertypes.Ballot, found bool)
 	GetAllBallots(ctx sdk.Context) (voters []*observertypes.Ballot)
 	GetParams(ctx sdk.Context) (params observertypes.Params)
-	GetCoreParamsByChainID(ctx sdk.Context, chainID int64) (params *observertypes.CoreParams, found bool)
+	GetChainParamsByChainID(ctx sdk.Context, chainID int64) (params *observertypes.ChainParams, found bool)
 	GetNodeAccount(ctx sdk.Context, address string) (nodeAccount observertypes.NodeAccount, found bool)
 	GetAllNodeAccount(ctx sdk.Context) (nodeAccounts []observertypes.NodeAccount)
 	SetNodeAccount(ctx sdk.Context, nodeAccount observertypes.NodeAccount)
@@ -91,6 +91,8 @@ type ObserverKeeper interface {
 	SetTssAndUpdateNonce(ctx sdk.Context, tss observertypes.TSS)
 	RemoveFromPendingNonces(ctx sdk.Context, tss string, chainID int64, nonce int64)
 	GetAllNonceToCctx(ctx sdk.Context) (list []observertypes.NonceToCctx)
+	GetSupportedChainFromChainID(ctx sdk.Context, chainID int64) *common.Chain
+	GetSupportedChains(ctx sdk.Context) []*common.Chain
 }
 
 type FungibleKeeper interface {