+++ x/observer/keeper/utils_test.go
@@ -5,16 +5,47 @@ import (
 	"testing"
 	"time"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	zetacommon "github.com/zeta-chain/zetacore/common"
 	keepertest "github.com/zeta-chain/zetacore/testutil/keeper"
 	"github.com/zeta-chain/zetacore/testutil/sample"
+	"github.com/zeta-chain/zetacore/x/observer/keeper"
 	"github.com/zeta-chain/zetacore/x/observer/types"
 )
+
+// setSupportedChain sets the supported chains for the observer module
+func setSupportedChain(ctx sdk.Context, observerKeeper keeper.Keeper, chainIDs ...int64) {
+	chainParamsList := make([]*types.ChainParams, len(chainIDs))
+	for i, chainID := range chainIDs {
+		chainParams := sample.ChainParams(chainID)
+		chainParams.IsSupported = true
+		chainParamsList[i] = chainParams
+	}
+	observerKeeper.SetChainParamsList(ctx, types.ChainParamsList{
+		ChainParams: chainParamsList,
+	})
+}
+
+// getValidEthChainIDWithIndex get a valid eth chain id with index
+func getValidEthChainIDWithIndex(t *testing.T, index int) int64 {
+	switch index {
+	case 0:
+		return zetacommon.GoerliLocalnetChain().ChainId
+	case 1:
+		return zetacommon.GoerliChain().ChainId
+	default:
+		require.Fail(t, "invalid index")
+	}
+	return 0
+}
 
 func TestKeeper_IsAuthorized(t *testing.T) {
 	t.Run("authorized observer", func(t *testing.T) {
 		k, ctx := keepertest.ObserverKeeper(t)
+
 		r := rand.New(rand.NewSource(9))
 
 		// Set validator in the store